test: extract master-detail-layout overlay tests to own suite
